test(TabSet): Split files
